+++ flutter_custom_tabs_ios/ios/flutter_custom_tabs_ios/Sources/flutter_custom_tabs_ios/Launcher.swift
@@ -31,7 +31,16 @@ extension Launcher {
 // MARK: - Impl
 
 final class DefaultLauncher: Launcher {
+  private let registrar: FlutterPluginRegistrar
   private var prewarmingTokenCache = [String: Any]()
+
+  private var viewController: UIViewController? {
+    registrar.viewController
+  }
+
+  init(registrar: FlutterPluginRegistrar) {
+    self.registrar = registrar
+  }
 
   func open(
     _ url: URL,
@@ -42,7 +51,10 @@ final class DefaultLauncher: Launcher {
   }
 
   func present(_ viewControllerToPresent: UIViewController, completion: ((Bool) -> Void)?) {
-    guard let topViewController = UIWindow.keyWindow?.topViewController() else {
+    guard
+      let viewController,
+      let topViewController = recursivelyFindTopViewController(from: viewController)
+    else {
       completion?(false)
       return
     }
@@ -52,7 +64,7 @@ final class DefaultLauncher: Launcher {
   }
 
   func dismissAll(completion: (() -> Void)?) {
-    guard let rootViewController = UIWindow.keyWindow?.rootViewController else {
+    guard let rootViewController = viewController else {
       completion?()
       return
     }
@@ -96,26 +108,6 @@ final class DefaultLauncher: Launcher {
     }
     token.invalidate()
     prewarmingTokenCache.removeValue(forKey: id)
-  }
-}
-
-private extension UIWindow {
-  static var keyWindow: UIWindow? {
-    // iOS 13+
-    if #available(iOS 13.0, *) {
-      return UIApplication.shared
-        .connectedScenes
-        .compactMap { $0 as? UIWindowScene }
-        .first { $0.activationState == .foregroundActive }?
-        .windows
-        .first(where: \ .isKeyWindow)
-    }
-    // iOS 12 fallback
-    return UIApplication.shared.windows.first(where: \ .isKeyWindow)
-  }
-
-  func topViewController() -> UIViewController? {
-    recursivelyFindTopViewController(from: rootViewController)
   }
 }